% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mode.R
\name{calc_mode}
\alias{calc_mode}
\title{Unimodal density estimator when the mode is unknown}
\usage{
calc_mode(x, t)
}
\arguments{
\item{x}{Vector of independent and identically distributed random variables;
must be sorted.}

\item{t}{A positive real number. Default value is one.}
}
\value{
\itemize{
\item mode - The estimator of the mode
\item x.knots - The vector of the knots of the estimated density.
\item F.knots - A vector consisting the values of the estimated
                distribution function evaluated at the knots.
\item f.knots - A vector whose i-th element gives the value of the
               estimated density on the segment
               joining the i-th and (i+1)-th knot.
               Recall that the estimated density is piecewise constant between two knots.
}
}
\description{
Estimates the density of a given sample under the assumption that
the underlying density is unimodal. The mode is estimated from
the data.
The method is based on the unimodal regularization of Birge (1997).
}
\details{
Birge(1997)'s  estimator gives a a pieciewise constant unimodal
density. The discontinuity points of the respective density are called the knots, which belong
to the set of datapoints. The density estimator is constant between two
consecutive knots.

 \code{t}:   The parameter \eqn{t}
                corresponds to the parameter \eqn{\tau}
                in Birge (1997). Higher values of \eqn{t} leads to more accurate
                 estimation of the unimodal densities.  This value ontrols the accuracy in
           unimodal density estimation upto the term \eqn{n^{-t}} where \eqn{n} is the sample
           size. We recommend a value greater than or equal to one.  See Birge (1997)
           for more details.
}
\examples{
x <- sort(rnorm(100)); calc_mode(x, 1)
}
\references{
Birge, L. (1997). \emph{Estimation of unimodal densities
         without smoothness assumptions}, Ann. Statist., 25, 970–981.
}
\author{
\href{https://connects.catalyst.harvard.edu/Profiles/display/Person/184207}{Nilanjana Laha}
(maintainer), \email{nlaha@hsph.harvard.edu},
Alex Luedtke, \email{aluedtke@uw.edu}.
}
