% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descrip.R
\name{do_description}
\alias{do_description}
\title{Descriptive statistics}
\usage{
do_description(d1, d2)
}
\arguments{
\item{d1}{first dataset}

\item{d2}{second dataset}
}
\value{
Makes plots of the ECDF, the Boxplot, the KDE, and the
   histograms of the datasets d1 and d2
}
\description{
Descriptive statistics
}
