% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd.birge.R
\name{hd.lc}
\alias{hd.lc}
\title{Hellinger distance between log-concave densities}
\usage{
hd.lc(x, y)
}
\arguments{
\item{x}{Vector of m independent and identically distributed random variables;
corresponds to the first sample.}

\item{y}{Vector of n independent and identically distributed random variables;
corresponds to the second sample.}
}
\value{
A point estimator of the Hellinger distance.
}
\description{
Provides an estimate of
the hellinger distance between two log-concave densities.
This function uses the log-concave density estimator of Dumbgen and Rufibatch (2009),
given by \link[logcondens]{logConDens} of logcondens package.
}
\examples{
x <- sort(rnorm(100)); y <- sort(rgamma(50, shape=1));
hd.lc(x,y)
}
\references{
Laha, N., Moodie, Z., Huang, Y., and Luedtke (2020), A.
\emph{ Improved inference for vaccine-induced immune responses
       via shape-constrained methods}. Submitted.

Dumbgen, L. and Rufibatch, K. (2009). \emph{Maximum likelihood estimation of a logconcave density and
          its distribution function: Basic properties and uniform
          consistency}, Bernoulli, 15, 40–68.
}
\seealso{
\code{\link{hd.uni}}, \code{\link{hd.lc.sm}}
}
\author{
\href{https://connects.catalyst.harvard.edu/Profiles/display/Person/184207}{Nilanjana Laha}
(maintainer), \email{nlaha@hsph.harvard.edu},

Alex Luedtke, \email{aluedtke@uw.edu}.
}
