% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd.birge.R
\name{hd.lc.sm}
\alias{hd.lc.sm}
\title{Hellinger distance between log-concave densities}
\usage{
hd.lc.sm(x, y)
}
\arguments{
\item{x}{Vector of m independent and identically distributed random variables;
corresponds to the first sample.}

\item{y}{Vector of n independent and identically distributed random variables;
corresponds to the second sample.}
}
\value{
A point estimator of the Hellinger distance.
}
\description{
Provides an estimate of
the squared Hellinger distance between two log-concave densities.
This function uses the smoothed log-concave density estimator of Chen and Samworth (2013),
given by \link[logcondens]{logConDens} of logcondens package.
}
\examples{
x <- sort(rnorm(100)); y <- sort(rgamma(50, shape=1));
hd.lc.sm(x,y)
}
\references{
Laha, N., Moodie, Z., Huang, Y., and Luedtke (2021), A.
\emph{ Improved inference for vaccine-induced immune responses
       via shape-constrained methods}. Submitted.

Chen, Y. and Samworth, R. J. (2013). \emph{Smoothed
          log-concave maximum likelihood estimation with applications},
          Statistica Sinica, 23, 1303-1398.
}
\seealso{
\code{\link{hd.lc}}, \code{\link{hd.uni}}, \code{\link{hell.ci}}
}
\author{
\href{https://connects.catalyst.harvard.edu/Profiles/display/Person/184207}{Nilanjana Laha}
(maintainer), \email{nlaha@hsph.harvard.edu},

Alex Luedtke, \email{aluedtke@uw.edu}.
}
