% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{hell.ci}
\alias{hell.ci}
\title{Confidence interval for Hellinger distance estimators}
\usage{
hell.ci(x, y, alpha = 0.05)
}
\arguments{
\item{x}{Vector of m independent and identically distributed random variables;
corresponds to the first sample.}

\item{y}{Vector of n independent and identically distributed random variables;
corresponds to the second sample.}

\item{alpha}{Level of significance; returns a \eqn{1-\alpha\%} confidence interval.}
}
\value{
\itemize{
\item lc.ci - Confidence intervals based on the log-concave MLE.
\item sm.lc.ci - Confidence intervals based on the smoothed log-concave MLE.
\item um.ci - Confidence intervals based on unimodal density estimator.
}
}
\description{
Calculates the confidence interval of the squared Hellinger distance, as
discussed in Laha et al. (2021). This function employs  shape
constrained methods to estimate the underlying densities. See below for more details.
which include the log-concave density estimator of Dumbgen and Rufibatch (2009),
and the smoothed log-concave density estimator
given by ,
}
\details{
The confidence intervals follow the construnction of Laha et al. (2021). One
 confidence interval is based on the unimodal density estimator of Birges (1997), and
 provably requires the underlying densities to be unimodal to work.
The other confidence intervals are based on the log-concave density estimators, i.e.
the log-concave MLE of Dumbgen and Rufibatch (2009), and the smoothed log-concave
MLE of Chen and Samworth (2013) .
Both are computed using the function \link[logcondens]{logConDens} of logcondens package.
}
\examples{
x <- sort(rnorm(100)); y <- sort(rgamma(50, shape=1));
hell.ci(x,y)$lc.ci
}
\references{
Laha, N., Moodie, Z., Huang, Y., and Luedtke (2020), A.
\emph{ Improved inference for vaccine-induced immune responses
       via shape-constrained methods}. Submitted.

Dumbgen, L. and Rufibatch, K. (2009). \emph{Maximum likelihood estimation of a logconcave density and
          its distribution function: Basic properties and uniform
          consistency}, Bernoulli, 15, 40–68.

Chen, Y. and Samworth, R. J. (2013). \emph{Smoothed
          log-concave maximum likelihood estimation with applications},
          Statistica Sinica, 23, 1303-1398.

Birge, L. (1997). \emph{Estimation of unimodal densities
         without smoothness assumptions}, Ann. Statist., 25, 970–981.
}
\seealso{
\code{\link{hd.uni}}, \code{\link{hd.lc.sm}}, \code{\link{hd.uni}}
}
\author{
\href{https://connects.catalyst.harvard.edu/Profiles/display/Person/184207}{Nilanjana Laha}
(maintainer), \email{nlaha@hsph.harvard.edu},

Alex Luedtke, \email{aluedtke@uw.edu}.
}
