% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{SDNN}
\alias{SDNN}
\title{One-sided test of  stochastic dominance against the null of non-dominance}
\usage{
SDNN(x, y, Method, t, p)
}
\arguments{
\item{x}{Vector of m independent and identically distributed random variables;
corresponds to the first sample.}

\item{y}{Vector of n independent and identically distributed random variables;
corresponds to the second sample.}

\item{Method}{Must be one among "NP" (nonparametric), "UM" (unimodal), and "LC" (log-concave). See 'Details'.}

\item{t}{A positive real number. Only required when Method={"UM"}, default value is 1.
See Details.}

\item{p}{The proportion of combined data to be trimmed from each end prior to testing,
should take value in \eqn{[0,0.50)},
the default is set to 0.05.}
}
\value{
A list of two numbers.
\itemize{
\item T1 - The p-value of the test based on minimum T-statistic of
           Laha \emph{et al.}, (2020)
\item T2 - The p-value of the TSEP test of Laha \emph{et al.}, (2020).
}
}
\description{
Calculates the p-values of one-sided
tests of restricted stochastic dominance against the null of non-dominance.
The concerned tests are the minimum t-statistic test
and the two sample empirical process (TSEP)
test of  \emph{Laha et al. (2020)}. Each test can be either nonparametric,
or semiparametric, ie.  using unimodality or log-concavity assumption on the underlying
densities.
}
\details{
Suppose \eqn{X_1,\ldots, X_m} and \eqn{Y_1,\ldots, y_n} are independent random
variables with distribution \eqn{F} and \eqn{G}, respectively. Denote by
\eqn{D_{p,m,n}} the set \deqn{[H_n^{-1}(p), H_n^{-1}(1-p)]} where
\eqn{p\in[0,0.5)} and \eqn{H_n} is the empirical distribution
function of the combined sample \deqn{\{X_1,\ldots, x_m,Y_1,\ldots,Y_n\}.} The function SDNN tests
 \eqn{H_0: F(x)\geq G(x)} for some \eqn{x\in D_{p,m,n}} vs
 \eqn{H_1: F(x)<G(x)} for all \eqn{x\in D_{p,m,n}}.
 For more details, see Laha et al., 2020.

\code{Method}:
            "NP" corresponds to the nonparametric tests. "UM" corresponds
             to tests which use the function \code{\link{calc_mode}}
             to estimate the densities of \eqn{X_i}'s and \eqn{Y_i}'s. This function
             estimates the unimodal density estimator of Birge (1997).
              "LC" corresponds to tests which use  the log-concave MLE (given by \link[logcondens]{logConDens} of R package ``logcondens")
              of Dumbgen and Rufibatch (2009) to estimate the latter densities. For more detail, see Laha et al. (2020).

\code{t}:   The parameter \eqn{t}
                corresponds to the parameter \eqn{\tau}
                in Birge (1997). Higher values of \eqn{t} leads to more accurate
                 estimation of the unimodal densities.  This value ontrols the accuracy in
           unimodal density estimation upto the term \eqn{(m+n)^{-t}}. A value greater
             than or equal to one is recommended.  See Birge (1997)
           for more details.

\code{p}:   p corresponds to the set \eqn{D_{p,m,n}} in \eqn{H_0} and \eqn{H_1}.
           To overcome the difficulties arising from the tail region,
          \eqn{100p} percent of data is trimmed from both sides of the
          combined sample.
}
\examples{
x <- rnorm(100); y <- rgamma(50, shape=1);
 SDNN(x, y, Method="UM", t=1, p=0.01)
}
\references{
Laha, N., Moodie, Z., Huang, Y., and Luedtke, A. (2020).
\emph{ Improved inference for vaccine-induced immune responses
       via shape-constrained methods}. Submitted.

Dumbgen, L. and Rufibatch, K. (2009). \emph{Maximum likelihood estimation of a logconcave density and
          its distribution function: Basic properties and uniform
          consistency}, Bernoulli, 15, 40–68.

Birge, L. (1997). \emph{Estimation of unimodal densities
         without smoothness assumptions}, Ann. Statist., 25, 970–981.
}
\seealso{
\code{\link{calc_mode}}, \link[logcondens]{logConDens}
}
\author{
\href{https://connects.catalyst.harvard.edu/Profiles/display/Person/184207}{Nilanjana Laha}
(maintainer), \email{nlaha@hsph.harvard.edu},
Alex Luedtke, \email{aluedtke@uw.edu}.
}
